requirements.txt
streamlit
beautifulsoup4
streamlit-lottie
GitPython